fixed return type of ReadEGG()

+++ MINT_EEGFirmware/MINT_EEG_Firmware.X/user.c
@@ -24,6 +24,6 @@ void InitJellyfish(void) {
 }
 
 
-float ReadEEG(void) {
+short ReadEEG(void) {
     return 0; 
 }
